Add kfp_support_lib tests to Travis.

Former-commit-id: 94b1a83da97747d2617fefe454b8fcaeec42ad76

+++ kind/hack/ingress.sh
@@ -6,12 +6,16 @@ source ${ROOT_DIR}/hack/common.sh
 
 deploy() {
 	kubectl apply -f ${ROOT_DIR}/hack/ray_api_server_ingress.yaml
+if [[ "${DEPLOY_KUBEFLOW}" -eq 1 ]]; then
 	kubectl apply -f ${ROOT_DIR}/hack/kfp_ingress.yaml
+fi
 }
 
 delete(){
 	kubectl delete -f ${ROOT_DIR}/hack/ray_api_server_ingress.yaml
+if [[ "${DEPLOY_KUBEFLOW}" -eq 1 ]]; then
 	kubectl delete -f ${ROOT_DIR}/hack/kfp_ingress.yaml
+fi
 }
 
 usage(){
